Changed float to double to avoid implicit double-promotion

+++ src/lib/parameters/param_translation.cpp
@@ -262,7 +262,7 @@ bool param_modify_on_import(bson_node_t node)
 	{
 		if (strcmp("VT_PSHER_RMP_DT", node->name) == 0) {
 			strcpy(node->name, "VT_PSHER_SLEW");
-			float _param_vt_f_trans_thr = param_find("VT_F_TRANS_THR");
+			double _param_vt_f_trans_thr = param_find("VT_F_TRANS_THR");
 			node->d = _param_vt_f_trans_thr / node->d;
 			PX4_INFO("copying %s -> %s", "VT_PSHER_RMP_DT", "VT_PSHER_SLEW");
 			return true;
